# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: termipod hub HTTP API
  description: |
    Authoritative OpenAPI specification for the termipod hub HTTP
    surface. Every `/v1/...` endpoint is listed by tag (resource
    group) with request / response shapes and the auth scheme.

    **Version model.** This spec covers `v1`. Additive changes
    (new endpoints, new optional fields) ship under `v1`; breaking
    changes would bump to `/v2/`. The hub reports its build at
    `GET /v1/_info` and the mobile client checks
    `supported_api_versions` at startup.

    **Cross-references.**
    - [`api-overview.md`](api-overview.md) — endpoint index +
      cross-cutting conventions (this spec is the machine-readable
      form).
    - [`database-schema.md`](database-schema.md) — physical schema
      that backs the response shapes.
    - [`audit-events.md`](audit-events.md) — the canonical taxonomy
      of `audit_events` rows referenced from response payloads.

    Status: draft. Hand-written; first issuance 2026-05-05 from
    `hub/internal/server/server.go` route table at v1.0.351.
  version: "1.0.351-alpha"
  license:
    name: Apache-2.0
    url: https://github.com/physercoe/termipod/blob/main/LICENSE
  contact:
    name: termipod issue tracker
    url: https://github.com/physercoe/termipod/issues
servers:
  - url: "https://hub.example.com"
    description: Production hub (replace with operator's hub URL)
  - url: "http://127.0.0.1:8443"
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/teams/{team}/references:
    parameters:
      - $ref: "#/components/parameters/Team"
    get:
      tags: [References]
      operationId: listReferences
      summary: List references in the team's library (ADR-053)
      parameters:
        - { name: collection, in: query, schema: { type: string } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: source, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          description: References
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Reference" }
    post:
      tags: [References]
      operationId: createReference
      summary: Add a reference to the library
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Reference" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reference" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/teams/{team}/references/{ref}:
    parameters:
      - $ref: "#/components/parameters/Team"
      - { name: ref, in: path, required: true, schema: { type: string } }
    get:
      tags: [References]
      operationId: getReference
      summary: Get one reference
      responses:
        "200":
          description: Reference
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reference" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [References]
      operationId: updateReference
      summary: Partially update a reference (present fields override)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Reference" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reference" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [References]
      operationId: deleteReference
      summary: Delete a reference
      responses:
        "200": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        SHA-256 hashed in `auth_tokens.token_hash`. Plaintext is
        printed only once at issuance. Token kinds: owner / user /
        agent / host. See `permission-model.md`.
  parameters:
    Team:
      in: path
      name: team
      required: true
      schema: { type: string }
      description: Team identifier (default `default` in MVP)
...
  responses:
    BadRequest:
      description: Malformed request (validation)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
...
  schemas:
    ProblemDetail:
      type: object
      description: RFC 7807 problem-detail
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        audit_event_id: { type: string }
        context:
          type: object
          additionalProperties: true
...
    Reference:
      type: object
      description: A research reference (ADR-053). Metadata only — PDF bytes are not stored here.
      properties:
        id: { type: string }
        team_id: { type: string }
        type:
          type: string
          enum: [article, preprint, book, report, webpage, note]
        title: { type: string }
        authors: { type: array, items: { type: string } }
        year: { type: integer, nullable: true }
        venue: { type: string }
        doi: { type: string }
        arxiv_id: { type: string }
        url: { type: string }
        pdf_url: { type: string }
        abstract: { type: string }
        tldr: { type: string }
        citation_count: { type: integer, nullable: true }
        source: { type: string }
        external_id: { type: string }
        tags: { type: array, items: { type: string } }
        collections: { type: array, items: { type: string } }
        notes: { type: string }
        body_markdown: { type: string }
        details:
          type: object
          additionalProperties: { type: string }
        zotero_storage:
          type: object
          properties:
            key: { type: string }
            file: { type: string }
            content_type: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...